+++ lib/dentaku/token_matchers.rb
@@ -6,7 +6,7 @@ module Dentaku
 
     def self.function_token_matchers(function_name, *symbols)
       token_matchers(:open, *symbols, :close).unshift(
-        TokenMatcher.send(function_name)
+        TokenMatcher.function(function_name)
       )
     end
 
@@ -16,9 +16,9 @@ module Dentaku
         :comparator, :comp_gt, :comp_lt, :open, :close, :comma,
         :non_close_plus, :non_group, :non_group_star, :arguments,
         :logical, :combinator, :if, :round, :roundup, :rounddown, :not,
-        :anchored_minus, :math_neg_pow, :math_neg_mul
+        :anchored_minus
       ].each_with_object({}) do |name, matchers|
-        matchers[name] = TokenMatcher.send(name)
+        matchers[name] = TokenMatcher.public_send(name)
       end
 
       @matchers.fetch(symbol) do